infosis-crowdmonitoring/requirements (1).txt
streamlit>=1.18.0
torch
torchvision
torchaudio
numpy
Pillow
opencv-python
matplotlib
scipy
scikit-learn
